==> scripts/misc/5_CalibrationFig.R <==
### LAST UPDATED 25 JUNE 2024 (v1.0).
### THIS SCRIPTS EXTRACTS THE CALIBRATION FIGURES FOR THE SUPPLEMENTARY NOTES.

.libPaths("H:/Programs/RLibrary")
library(dplyr)
library(readr)
library(stringr)
library(ggplot2)
library(gridExtra)

read_tsv("data/output/res/")

#! Fix this one...
